refactor: extract portal_layouts http feature into ferriskey-api-portal-layouts
