+++ libs/ferriskey-api-portal-layouts/src/handlers/update_layout.rs
@@ -6,58 +6,60 @@ use ferriskey_core::domain::{
     authentication::value_objects::Identity,
     portal_layouts::{
         entities::PortalLayout,
-        ports::{CreateLayoutInput, PortalLayoutsService},
+        ports::{PortalLayoutsService, UpdateLayoutInput},
     },
 };
 use serde::{Deserialize, Serialize};
 use utoipa::ToSchema;
+use uuid::Uuid;
 
-use crate::application::http::{
-    portal_layouts::validators::CreatePortalLayoutValidator,
-    server::{
-        api_entities::{
-            api_error::{ApiError, ApiErrorResponse, ValidateJson},
-            response::Response,
-        },
-        app_state::AppState,
+use crate::validators::UpdatePortalLayoutValidator;
+use ferriskey_api_core::{
+    api_entities::{
+        api_error::{ApiError, ApiErrorResponse, ValidateJson},
+        response::Response,
     },
+    app_state::AppState,
 };
 
 #[derive(Debug, Serialize, Deserialize, ToSchema)]
-pub struct CreatePortalLayoutResponse {
+pub struct UpdatePortalLayoutResponse {
     pub data: PortalLayout,
 }
 
 #[utoipa::path(
-    post,
-    path = "",
+    put,
+    path = "/{layout_id}",
     tag = "portal-layouts",
-    summary = "Create a portal layout",
-    description = "Creates a new portal layout. The first layout in a realm is automatically marked as default.",
+    summary = "Update a portal layout",
+    description = "Updates the name and tree of an existing portal layout.",
     params(
         ("realm_name" = String, Path, description = "Name of the realm"),
+        ("layout_id" = Uuid, Path, description = "Portal layout ID"),
     ),
-    request_body = CreatePortalLayoutValidator,
+    request_body = UpdatePortalLayoutValidator,
     responses(
-        (status = 201, description = "Layout created successfully", body = CreatePortalLayoutResponse),
+        (status = 200, description = "Layout updated successfully", body = UpdatePortalLayoutResponse),
         (status = 400, description = "Invalid request data", body = ApiErrorResponse),
+        (status = 404, description = "Layout not found", body = ApiErrorResponse),
         (status = 401, description = "Unauthorized", body = ApiErrorResponse),
         (status = 403, description = "Insufficient permissions", body = ApiErrorResponse),
         (status = 500, description = "Internal server error", body = ApiErrorResponse),
     ),
 )]
-pub async fn create_layout(
-    Path(realm_name): Path<String>,
+pub async fn update_layout(
+    Path((realm_name, layout_id)): Path<(String, Uuid)>,
     State(state): State<AppState>,
     Extension(identity): Extension<Identity>,
-    ValidateJson(payload): ValidateJson<CreatePortalLayoutValidator>,
-) -> Result<Response<CreatePortalLayoutResponse>, ApiError> {
+    ValidateJson(payload): ValidateJson<UpdatePortalLayoutValidator>,
+) -> Result<Response<UpdatePortalLayoutResponse>, ApiError> {
     let layout = state
         .service
-        .create_layout(
+        .update_layout(
             identity,
-            CreateLayoutInput {
+            UpdateLayoutInput {
                 realm_name,
+                layout_id,
                 name: payload.name,
                 tree: payload.tree,
             },
@@ -65,7 +67,7 @@ pub async fn create_layout(
         .await
         .map_err(ApiError::from)?;
 
-    Ok(Response::Created(CreatePortalLayoutResponse {
+    Ok(Response::Updated(UpdatePortalLayoutResponse {
         data: layout,
     }))
 }
